django_blog/requirements.txt
Django>=4.2
django-taggit>=3.0